refactor: PathUtils 역할 Path에게 위임
